+++ globals.js
@@ -22,6 +22,8 @@ module.exports = (env = { TARGET: 'sdk' }) => ({
 
     __MESSAGES__: {
         __VERSION__: env.VERSION || version,
+        __FEATURES__: 'no-common',
+        __DEMO__: !!env.demo,
         __TARGET__: env.TARGET.toUpperCase(),
         __STAGE_TAG__: env.STAGE_TAG,
         __TEST_ENV__: env.TEST_ENV,