+++ src/components/modal/parts/Container.jsx
@@ -1,5 +1,4 @@
 /** @jsx h */
-import stringStartsWith from 'core-js-pure/stable/string/starts-with';
 import { h } from 'preact';
 import { useEffect, useState } from 'preact/hooks';
 import { getOrCreateStorageID } from '../../../utils';
@@ -14,11 +13,11 @@ const Container = ({ children, contentWrapper, contentMaxWidth, contentMaxHeight
     const [loading, setLoading] = useState(false);
 
     useEffect(() => {
-        if (transitionState === 'OPENING') {
-            window.focus();
-
+        if (transitionState === 'CLOSED') {
             // eslint-disable-next-line no-param-reassign
             contentWrapper.current.scrollTop = 0;
+        } else if (transitionState === 'OPEN') {
+            window.focus();
         }
     }, [transitionState]);
 
@@ -37,7 +36,7 @@ const Container = ({ children, contentWrapper, contentMaxWidth, contentMaxHeight
         setLoading(true);
         getContent({
             currency,
-            amount,
+            amount: amount === '' ? undefined : amount,
             payerId,
             clientId,
             merchantId,
@@ -53,11 +52,7 @@ const Container = ({ children, contentWrapper, contentMaxWidth, contentMaxHeight
     return (
         <ScrollProvider containerRef={contentWrapper}>
             <div className="modal-wrapper">
-                <section
-                    className={`modal-container ${stringStartsWith(transitionState, 'OPEN') ? 'show' : ''} ${
-                        loading ? 'loading' : ''
-                    }`}
-                >
+                <section className={`modal-container show ${loading ? 'loading' : ''}`}>
                     <div className="spinner" style={{ opacity: loading ? '1' : '0' }} />
                     <div className="wrapper">{children}</div>
                     <Overlay contentMaxWidth={contentMaxWidth} contentMaxHeight={contentMaxHeight} />