+++ tests/functional/spec/DE/modal/deInst_modalTestDefs.js
@@ -56,7 +56,7 @@ export const updateFinanceTermsInst = ({ account, viewport, groupString }) => as
 
     await modalFrame.waitForSelector(selectors.calculator.calc, { visible: true });
     await modalFrame.click(selectors.calculator.calcInput, { clickCount: 3 });
-    await modalFrame.type(selectors.calculator.calcInput, '650');
+    await modalFrame.type(selectors.calculator.calcInput, '650', { delay: 100 });
     await modalFrame.click(selectors.button.btnMd);
     await page.waitFor(4 * 1000);