+++ tests/functional/spec/utils/setupTestPage.js
@@ -1,20 +1,26 @@
 import selectors from './selectors';
 
-export default async function setTestPage({ config, testPage, frameName }) {
+export default async function setupTestPage({ config, testPage, frameName }) {
     await page.goto(`https://localhost.paypal.com:8080/snapshot/${testPage}?config=${JSON.stringify(config)}`);
 
     const frameWithMessage = frameName ? page.frames().find(frame => frame.name() === frameName) : page.mainFrame();
 
-    await frameWithMessage.waitForSelector(selectors.banner.iframeByAttribute, { visible: true });
-    await frameWithMessage.waitForSelector(selectors.modal.iframe);
-
-    const bannerElement = await frameWithMessage.$(selectors.banner.iframeByAttribute);
-    const modalElement = await frameWithMessage.$(selectors.modal.iframe);
+    const bannerElement = await frameWithMessage.waitForSelector(selectors.banner.iframeByAttribute, { visible: true });
     const bannerFrame = await bannerElement.contentFrame();
-    const modalFrame = await modalElement.contentFrame();
 
-    await modalFrame.waitForSelector(selectors.modal.contentBody);
     await bannerFrame.waitForSelector(selectors.banner.messageMessaging, { visible: true });
 
-    return { bannerFrame, modalFrame };
+    const openModal = async () => {
+        // Modal is no longer preloaded so must click on the banner to render it
+        await bannerElement.click();
+
+        const modalElement = await page.waitForSelector(selectors.modal.iframe, { visible: true });
+        const modalFrame = await modalElement.contentFrame();
+
+        await modalFrame.waitForSelector(selectors.modal.contentBody);
+
+        return { modalFrame };
+    };
+
+    return { bannerFrame, openModal };
 }